Add a read checksum policy for ImageFile() (API change)
E57 files are read in blocks of 1 KB - each with a checksum. Calculating and verifying the
checksum on every block is very expensive and makes reading of large files very slow
(checks every 1 KB on a 124 GB file for example). The developer can now specify a
checksum policy to sacrifice verification for speed.

Checksum policy is set as a percent of blocks to check (0-100). Some convenience
consts are defined:

  CHECKSUM_POLICY_NONE - do not compute and verify
  CHECKSUM_POLICY_SPARSE - verify 25% of blocks
  CHECKSUM_POLICY_HALF - verify 50% of blocks
  CHECKSUM_POLICY_ALL - verify all blocks (default)

Except in the case of CHECKSUM_POLICY_NONE, the last block is always checked.