E57Simple: Fix min/max fields in Data3D
There was a mixture of using float and double min/max.

When using float, they should be set to E57_FLOAT_MIN and E57_FLOAT_MAX.
When using double, they should be set to E57_DOUBLE_MIN and E57_DOUBLE_MAX.

Do this adjustment to Data3D automatically in the Data3DPointsData_t() constructor and add tests for it.