{doc} Remove confusing statement in StructureNode and VectorNode isDefined()
"If this (Structure|Vector)Node is not attached to an ImageFile, the pathName origin root will not the root node of an ImageFile."

I think it's missing "be"—"will not be the root node".

However even then it seems to be a useless statement: "if it's not attached to an image file, the root won't be from an image file".

[skip ci]